+++ include/slang/ast/expressions/OperatorExpressions.h
@@ -259,18 +259,18 @@ public:
         std::optional<bitwidth_t> constantWithWidth;
     };
 
-    StreamingConcatenationExpression(const Type& type, uint32_t sliceSize, uint32_t bitstreamWidth,
+    StreamingConcatenationExpression(const Type& type, uint64_t sliceSize, uint64_t bitstreamWidth,
                                      std::span<const StreamExpression> streams,
                                      SourceRange sourceRange) :
         Expression(ExpressionKind::Streaming, type, sourceRange),
         streams_(streams), sliceSize(sliceSize), bitstreamWidth(bitstreamWidth) {}
 
     bool isFixedSize() const;
-    uint32_t getBitstreamWidth() const { return bitstreamWidth; }
+    uint64_t getBitstreamWidth() const { return bitstreamWidth; }
 
     /// Gets the size of the blocks to slice and reorder: if 0, this is a left-to-right
     /// concatenation. Otherwise, it's a right-to-left concatenation.
-    uint32_t getSliceSize() const { return sliceSize; }
+    uint64_t getSliceSize() const { return sliceSize; }
 
     std::span<const StreamExpression> streams() const { return streams_; }
 
@@ -295,8 +295,8 @@ public:
 
 private:
     std::span<const StreamExpression> streams_;
-    uint32_t sliceSize;
-    uint32_t bitstreamWidth;
+    uint64_t sliceSize;
+    uint64_t bitstreamWidth;
 };
 
 /// Denotes a range of values by providing expressions for the lower and upper
